Split notification mute support files
