+++ Sources/NotificationDisplaySettings.swift
@@ -1,0 +1,30 @@
+import Foundation
+
+enum NotificationBadgeSettings {
+    static let dockBadgeEnabledKey = "notificationDockBadgeEnabled"
+    static let defaultDockBadgeEnabled = true
+
+    static func isDockBadgeEnabled(defaults: UserDefaults = .standard) -> Bool {
+        if defaults.object(forKey: dockBadgeEnabledKey) == nil {
+            return defaultDockBadgeEnabled
+        }
+        return defaults.bool(forKey: dockBadgeEnabledKey)
+    }
+}
+
+enum NotificationPaneRingSettings {
+    static let enabledKey = "notificationPaneRingEnabled"
+    static let defaultEnabled = true
+}
+
+enum NotificationPaneFlashSettings {
+    static let enabledKey = "notificationPaneFlashEnabled"
+    static let defaultEnabled = true
+
+    static func isEnabled(defaults: UserDefaults = .standard) -> Bool {
+        if defaults.object(forKey: enabledKey) == nil {
+            return defaultEnabled
+        }
+        return defaults.bool(forKey: enabledKey)
+    }
+}
